# docs/swagger.yaml
basePath: /api
...
info:
  contact:
    email: support@soundcave.com
    name: API Support
  description: API documentation for SoundCave Backend - Music streaming platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SoundCave Backend API
  version: "1.0"
paths:
...
  /dashboard/customer-report:
    get:
      consumes:
      - application/json
      description: Get customer report with revenue, growth, and analytics
      parameters:
      - default: 30
        description: Report period in days
        in: query
        name: period
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get customer report
      tags:
      - Dashboard
  /dashboard/stats:
    get:
      consumes:
      - application/json
      description: Get dashboard statistics including counts and metrics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard statistics
      tags:
      - Dashboard
...
securityDefinitions:
  BearerAuth:
    description: 'To authenticate, provide the API token with the ''Bearer '' prefix.
      Example: ''Bearer your_token_here'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"